requirements.txt
google-genai>=2.0.0
supabase>=2.10.0
aiohttp>=3.9.0
httpx>=0.27.0
python-dotenv>=1.0.0
